Factor LC provenance preservation helper
